User Route Authorization

+++ app/Http/Controllers/MovieController.php
@@ -12,9 +12,15 @@ class MovieController extends Controller
 {
   //Using Form Request to peform validations while resources and collections to return response to improve code readability.
 
+  public function __construct()
+    {
+        $this->middleware('auth:sanctum');
+    }
+
     //Show all movies
     public function index()
     {
+        
         return  new MovieCollection(Movie::all());
     }
 
@@ -27,6 +33,7 @@ class MovieController extends Controller
             'genre' => $request->input('genre'),
             'producer' => $request->input('producer'),
             'synopsis' => $request->input('synopsis'),
+            'user_id'=>auth()->user()->id,
         ]);
 
         return new MovieResource($movie);
@@ -36,12 +43,22 @@ class MovieController extends Controller
     //Get a particular movie
     public function show(Movie $movie)
     {
+        $user_id = auth()->user()->id;
+        $check = Movie::where('user_id',$user_id)->get();
+        if(count($check)==0) {
+            return response()->json(['message'=>'Document not found'],404);
+        }
         return new MovieResource($movie);
     }
 
   //Update a Movie
     public function update(MovieRequest $request, Movie $movie)
     {
+        $user_id = auth()->user()->id;
+        $check = Movie::where('user_id',$user_id)->get();
+        if(count($check)==0) {
+            return response()->json(['message'=>'Document not found'],404);
+        }
      $movie->update($request->all());
      return new MovieResource($movie);
     }
@@ -50,6 +67,11 @@ class MovieController extends Controller
     //Delete a movie
     public function destroy(Movie $movie)
     {
+        $user_id = auth()->user()->id;
+        $check = Movie::where('user_id',$user_id)->get();
+        if(count($check)==0) {
+            return response()->json(['message'=>'Document not found'],404);
+        }
         $movie->delete();
         return response()->json(['movie'=>'Movie Deleted'], 200);
