+++ app/Http/Controllers/RentalController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\RentalRequest;
 use App\Http\Resources\Rental\Rental as RentalResource;
 use App\Http\Resources\Rental\RentalCollection;
 use App\Models\Rental;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 
@@ -14,6 +15,12 @@ class RentalController extends Controller
 {
   //Using Form Request to peform validations while resources and collections to return response to improve code readability.
 
+  public function __construct()
+  {
+      $this->middleware('auth:sanctum');
+  }
+
+   
 
   //Show all Rentals
     public function index()
@@ -24,9 +31,12 @@ class RentalController extends Controller
         //Create A Rental
     public function store(RentalRequest $request)
     {
+        $user_id = auth()->user()->id;
         $rental = Rental::create([
             'movie_title' => $request->input('movie_title'),
             'datetime' => $request->input('datetime'),
+            'user_id'=>$user_id,
+
           
         ]);
 
@@ -37,12 +47,22 @@ class RentalController extends Controller
     public function show(Rental $rental)
     {
         
+        $user_id = auth()->user()->id;
+        $check = Rental::where('user_id',$user_id)->get();
+        if(count($check)==0) {
+            return response()->json(['message'=>'Document not found'],404);
+        }
         return new RentalResource($rental);
     }
 
     //Update A Rental
     public function update(RentalRequest $request, Rental $rental)
     {
+        $user_id = auth()->user()->id;
+        $check = Rental::where('user_id',$user_id)->get();
+        if(count($check)==0) {
+            return response()->json(['message'=>'Document not found'],404);
+        }
      $rental->update($request->all());
      return new RentalResource($rental);
     }
@@ -50,6 +70,11 @@ class RentalController extends Controller
     //Delete A Rental
         public function destroy(Rental $rental)
     {
+        $user_id = auth()->user()->id;
+        $check = Rental::where('user_id',$user_id)->get();
+        if(count($check)==0) {
+            return response()->json(['message'=>'Document not found'],404);
+        }
         $rental->delete();
         return response()->json(['rental'=>'Rental Deleted'], 200);
